more -Wconversion issues
git-svn-id: https://svn.r-project.org/R/trunk@59184 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/src/bandwidths.c
@@ -112,9 +112,9 @@ band_den_bin(int *n, int *nb, double *d, double *x, int *cnt)
     rang = (xmax - xmin) * 1.01;
     *d = dd = rang / (*nb);
     for (i = 1; i < nn; i++) {
-	ii = x[i] / dd;
+	ii = (int)(x[i] / dd);
 	for (j = 0; j < i; j++) {
-	    jj = x[j] / dd;
+	    jj = (int)(x[j] / dd);
 	    iij = abs9((ii - jj));
 	    cnt[iij]++;
 	}
