+++ src/library/stats/src/d2x2xk.c
@@ -14,8 +14,8 @@ d2x2xk(Sint *k, double *m, double *n, double *t, double *d)
     c[0] = (double *) R_alloc(1, sizeof(double));
     c[0][0] = 1;
     for(i = 0; i < *k; i++) {
-	y = imax2(0, *t - *n);
-	z = imin2(*m, *t);
+	y = imax2(0,  (int)(*t - *n));
+	z = imin2((int)*m, (int)*t);
 	c[i + 1] = (double *) R_alloc(l + z - y + 1, sizeof(double));
 	for(j = 0; j <= l + z - y; j++)
 	    c[i + 1][j] = 0;
